CMake: Setup test projects.

+++ examples/rotateModel/renderSystem.cpp
@@ -1,6 +1,6 @@
 #include "renderSystem.hpp"
-#include "swapchain.hpp"
-#include "vertex.hpp"
+#include "lwEngine/swapchain.hpp"
+#include "lwEngine/vertex.hpp"
 
 #define GLM_FORCE_RADIANS
 #include <glm/glm.hpp>
@@ -68,7 +68,7 @@ void RenderSystem::updateUniformBuffer(uint32_t currentImage, VkExtent2D frameEx
 	UniformBufferObject ubo{};
 	ubo.model = glm::rotate(glm::mat4(1.0f), time * glm::radians(90.0f), glm::vec3(0.0f, 0.0f, 1.0f));
 	ubo.view = glm::lookAt(glm::vec3(2.0f, 2.0f, 2.0f), glm::vec3(0.0f, 0.0f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f));
-	ubo.proj = glm::perspective(glm::radians(45.0f), frameExtent.width / static_cast<float>(frameExtent.height), 0.1f, 10.0f);
+	ubo.proj = glm::perspective(glm::radians(45.0f), static_cast<float>(frameExtent.width) / static_cast<float>(frameExtent.height), 0.1f, 10.0f);
 	ubo.proj[1][1] *= -1;  // Invert the y-coordinate of clip coordinate because glm was designed for OpenGL
 	ubo.offset = offset;
